+++ src/Editor/index.js
@@ -82,6 +82,7 @@ export default class WysiwygEditor extends Component {
     customBlockRenderFunc: PropTypes.func,
     wrapperId: PropTypes.number,
     customDecorators: PropTypes.array,
+    editorRef: PropTypes.func,
   };
 
   static defaultProps = {
@@ -234,8 +235,8 @@ export default class WysiwygEditor extends Component {
   };
 
   setEditorReference: Function = (ref: Object): void => {
-    if (this.props.ref) {
-      this.props.ref(ref);
+    if (this.props.editorRef) {
+      this.props.editorRef(ref);
     }
     this.editor = ref;
   };
@@ -317,7 +318,7 @@ export default class WysiwygEditor extends Component {
   }
 
   filterEditorProps = props => filter(props, [
-    'ref', 'onChange', 'onEditorStateChange', 'onContentStateChange', 'initialContentState',
+    'onChange', 'onEditorStateChange', 'onContentStateChange', 'initialContentState',
     'defaultContentState', 'contentState', 'editorState', 'defaultEditorState', 'locale',
     'localization', 'toolbarOnFocus', 'toolbar', 'toolbarCustomButtons', 'toolbarClassName',
     'editorClassName', 'toolbarHidden', 'wrapperClassName', 'toolbarStyle', 'editorStyle',